+++ packages/sitecore-jss-angular/.eslintrc
@@ -4,7 +4,6 @@
     "../../.eslintrc"
 	],
 	"rules": {
-		"@typescript-eslint/no-explicit-any": "off",
 		"@typescript-eslint/explicit-module-boundary-types": "off"
 	}
 }
