+++ packages/sitecore-jss-angular/src/components/rendering-field.ts
@@ -1,6 +1,6 @@
 /* eslint-disable @typescript-eslint/no-empty-interface */
-export interface RenderingField {
-  value?: { [key: string]: unknown };
+export interface RenderingField<V = unknown> {
+  value?: V;
   editable?: string;
 }
 
@@ -9,10 +9,14 @@ export interface DateField {
   editable?: string;
 }
 
-export interface FileField extends RenderingField {
+export interface FileFieldValue {
   src?: string;
   title?: string;
   displayName?: string;
+}
+
+export interface FileField extends FileFieldValue, RenderingField {
+  value?: FileFieldValue;
 }
 
 export interface ImageFieldValue {
@@ -39,6 +43,6 @@ export interface LinkField extends LinkFieldValue, RenderingField {
   editableLastPart?: string;
 }
 
-export interface RichTextField extends RenderingField {}
+export interface RichTextField extends RenderingField<string> {}
 
-export interface TextField extends RenderingField {}
+export interface TextField extends RenderingField<string> {}
