+++ packages/sitecore-jss-angular/src/components/text.directive.ts
@@ -38,11 +38,7 @@ export class TextDirective implements OnChanges {
     const field = this.field;
     let editable = this.editable;
 
-    if (
-      !field ||
-      (!field.editable &&
-        (field.value === undefined || (typeof field.value === 'string' && field.value === '')))
-    ) {
+    if (!field || (!field.editable && (field.value === undefined || field.value === ''))) {
       return;
     }
