+++ packages/sitecore-jss-angular/src/jss-component-factory.service.ts
@@ -19,8 +19,10 @@ import { RawComponent } from './components/raw.component';
 import { isRawRendering } from './components/rendering';
 
 export interface ComponentFactoryResult {
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
   componentImplementation?: Type<any>;
   componentDefinition: ComponentRendering | HtmlElementRendering;
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
   componentFactory?: ComponentFactory<any>;
 }
 
@@ -81,6 +83,7 @@ export class JssComponentFactoryService {
         }
 
         if (component.componentName in dynamicComponentType) {
+          // eslint-disable-next-line @typescript-eslint/no-explicit-any
           componentType = (dynamicComponentType as { [s: string]: any })[component.componentName];
         } else {
           if (typeof dynamicComponentType === 'function') {
