+++ packages/sitecore-jss-nextjs/global.d.ts
@@ -10,3 +10,14 @@ declare module 'sync-disk-cache' {
     remove(key: string): void;
   }
 }
+
+declare namespace NodeJS {
+  export interface Global {
+    [key: string]: unknown;
+    requestAnimationFrame: (callback: () => void) => void;
+    window: Window;
+    document: Document;
+    navigator: Navigator;
+    HTMLElement: HTMLElement;
+  }
+}
