+++ packages/sitecore-jss-nextjs/src/tests/shim.ts
@@ -1,11 +1,12 @@
+/* eslint-disable spaced-comment */
 // React 16 depends on requestAnimationFrame, need a shim for node.js
 // eslint-disable-next-line spaced-comment
 // https://github.com/facebook/jest/issues/4545
 
-import { Global } from './jsdom-setup';
+/// <reference types="../../global" />
 
 // eslint-disable-next-line no-var
-declare var global: Global;
+declare var global: NodeJS.Global;
 
 global.requestAnimationFrame = (callback: () => void) => {
   setTimeout(callback, 0);
