+++ packages/sitecore-jss-nextjs/tsconfig.json
@@ -14,6 +14,7 @@
     "noImplicitReturns": true,
     "noUnusedLocals": true,
     "noUnusedParameters": true,
+    "noImplicitAny": true,
     "strict": true,
     "typeRoots": ["node_modules/@types"],
     "outDir": "dist/cjs",
