+++ packages/sitecore-jss-react-forms/global.d.ts
@@ -1,1 +1,12 @@
 declare module 'style-attr';
+
+declare namespace NodeJS {
+  export interface Global {
+    [key: string]: unknown;
+    requestAnimationFrame: (callback: () => void) => void;
+    window: Window;
+    document: Document;
+    navigator: Navigator;
+    HTMLElement: HTMLElement;
+  }
+}
