+++ packages/sitecore-jss-react-forms/src/tests/shim.ts
@@ -1,9 +1,10 @@
+/* eslint-disable spaced-comment */
 // React 16 depends on requestAnimationFrame, need a shim for node.js
 // https://github.com/facebook/jest/issues/4545
 
-import { Global } from './jsdom-setup';
+/// <reference types="../../global" />
 
-declare let global: Global;
+declare let global: NodeJS.Global;
 
 global.requestAnimationFrame = (callback) => {
   setTimeout(callback, 0);
