+++ packages/sitecore-jss-react/src/index.ts
@@ -27,8 +27,8 @@ export {
 export { ComponentFactory } from './components/sharedTypes';
 export { Placeholder } from './components/Placeholder';
 export { Image, ImageField } from './components/Image';
-export { RichText, RichTextProps, RichTextPropTypes } from './components/RichText';
-export { Text } from './components/Text';
+export { RichText, RichTextProps, RichTextPropTypes, RichTextField } from './components/RichText';
+export { Text, TextField } from './components/Text';
 export { DateField } from './components/Date';
 export { Link, LinkField, LinkFieldValue, LinkProps, LinkPropTypes } from './components/Link';
 export { File, FileField } from './components/File';
