+++ packages/sitecore-jss-rendering-host/tsconfig.json
@@ -11,6 +11,7 @@
     "noFallthroughCasesInSwitch": true,
     "noImplicitReturns": true,
     "noUnusedLocals": true,
+    "noImplicitAny": true,
     "noUnusedParameters": false,
     "strict": true,
     "typeRoots": ["node_modules/@types"],
